Update `appinsights` to released version (#2033)
I noticed that there was a note to update `appinsights` to a released version once one was available, and now one is.

The `close` function was changed to be `async` in the latest version, so `try_flush_and_close` must also become async. Luckily, the three places we use this already have an async context available.